fix: restore unit struct handling
it doesn't make much sense for the derive to do this, but it used to be
supported, and I can't really see why it should be dropped. if it should
be dropped, I think we should work on getting a better error message
than just panicing produces.

also adds a test case for this, misplacing it under the tuple_struct
file.